Allow running jobs to be stopped
This reimplements resque-status' kill functionality. Using the verb
"kill" always seemed like a misnomer since it will not kill a stuck job.

Instead we've used the verb "stop" since that's what we're asking the
job to do. We've taken some liberties by using the "reserved" field
"stop" to implement this functionality. It seemed like a natural fit.

Raising `Stopped` also felt like a natural fit. It means the job will
not be retried, which is what we want here.